+++ src/parser/toMarkup.ts
@@ -7,7 +7,9 @@ import parseDoc from './parseDoc';
 export function toMarkup(template: string): [Markup, Spaces] {
     // Replace out of date markers before parsing
     const tokens = toTokens(
-        DOCS_SYMBOL + template.replaceAll('$!', '').trim() + DOCS_SYMBOL
+        (template.startsWith(DOCS_SYMBOL) ? '' : DOCS_SYMBOL) +
+            template.replaceAll('$!', '').trim() +
+            (template.endsWith(DOCS_SYMBOL) ? '' : DOCS_SYMBOL),
     );
     return [parseDoc(tokens).markup, tokens.getSpaces()];
 }